# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Moments API",
        "contact": {},
        "version": "0.2.1"
    },
    "paths": {
...
        "/api/sysConfig/get": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysConfig"
                ],
                "summary": "获取系统设置(完整的)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sysConfig/save": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysConfig"
                ],
                "summary": "保存系统设置",
                "parameters": [
                    {
                        "description": "保存系统设置",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.FullSysConfigVO"
                        }
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.loginSuccessDTO"
                        }
                    }
                }
            }
        },
        "/api/user/profile/{username}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取指定用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "string",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.User"
                        }
                    }
                }
            }
        },
        "/api/user/reg": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.RegReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/saveProfile": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "保存用户信息",
                "parameters": [
                    {
                        "description": "保存用户信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ProfileReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Comment": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "$ref": "#/definitions/db.Memo"
                },
                "memoId": {
                    "type": "integer"
                },
                "replyTo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "db.Memo": {
            "type": "object",
            "properties": {
                "bilibiliUrl": {
                    "type": "string"
                },
                "commentCount": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Comment"
                    }
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "externalFavicon": {
                    "type": "string"
                },
                "externalTitle": {
                    "type": "string"
                },
                "externalUrl": {
                    "type": "string"
                },
                "favCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imgs": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "music163Url": {
                    "type": "string"
                },
                "pinned": {
                    "type": "boolean"
                },
                "showType": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.User"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "db.User": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "type": "string"
                },
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "beianNo": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "coverUrl": {
                    "description": "封面URL",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "css": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "enableS3": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "favicon": {
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "js": {
                    "type": "string"
                },
                "memos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Memo"
                    }
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                },
                "slogan": {
                    "description": "标语",
                    "type": "string"
                },
                "thumbnailSuffix": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "handler.loginSuccessDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "token": {
                    "description": "token",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "vo.FullSysConfigVO": {
            "type": "object",
            "properties": {
                "adminUserName": {
                    "description": "管理员名称",
                    "type": "string"
                },
                "beiAnNo": {
                    "description": "备案号码",
                    "type": "string"
                },
                "commentOrder": {
                    "description": "评论展示的顺序,asc:顺序,desc:逆序",
                    "type": "string"
                },
                "css": {
                    "description": "自定义css",
                    "type": "string"
                },
                "enableComment": {
                    "description": "是否启用评论",
                    "type": "boolean"
                },
                "enableGoogleRecaptcha": {
                    "description": "是否启用google recaptcha",
                    "type": "boolean"
                },
                "enableRegister": {
                    "description": "是否开启注册用户",
                    "type": "boolean"
                },
                "enableS3": {
                    "description": "是否启用S3",
                    "type": "boolean"
                },
                "favicon": {
                    "description": "favicon",
                    "type": "string"
                },
                "googleSecretKey": {
                    "description": "google recaptcha secretKey",
                    "type": "string"
                },
                "googleSiteKey": {
                    "description": "google recaptcha siteKey",
                    "type": "string"
                },
                "js": {
                    "description": "自定义js",
                    "type": "string"
                },
                "maxCommentLength": {
                    "description": "发言最大长度",
                    "type": "integer"
                },
                "memoMaxHeight": {
                    "description": "单个memo的最大高度,单位px",
                    "type": "integer"
                },
                "s3": {
                    "description": "S3相关信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.S3VO"
                        }
                    ]
                },
                "timeFormat": {
                    "description": "时间格式",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
...
        "vo.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "vo.ProfileReq": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "coverUrl": {
                    "description": "封面URL",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码,不修改不要填写",
                    "type": "string"
                },
                "slogan": {
                    "description": "标语",
                    "type": "string"
                }
            }
        },
        "vo.RegReq": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "repeatPassword": {
                    "description": "重复密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "vo.S3VO": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "description": "AK",
                    "type": "string"
                },
                "bucket": {
                    "description": "桶名称",
                    "type": "string"
                },
                "domain": {
                    "description": "S3 域名",
                    "type": "string"
                },
                "endpoint": {
                    "description": "S3接口地址",
                    "type": "string"
                },
                "region": {
                    "description": "地区",
                    "type": "string"
                },
                "secretKey": {
                    "description": "SK",
                    "type": "string"
                },
                "thumbnailSuffix": {
                    "description": "图片后缀",
                    "type": "string"
                }
            }
        },
...
    }
}